pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentbench"
version = "0.1.0"
description = "Standardized benchmark framework for comparing AI coding agents"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "radon>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
agentbench = "agentbench.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agentbench"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["agentbench"]

[tool.coverage.report]
show_missing = true
fail_under = 80
